@@ -16,7 +16,8 @@ const httpOptions = {
 })
 export class SnippetService {
 
-  private apiUrl = 'http://localhost:8080/api/snippet'
+  //private apiUrl = 'http://localhost:8080/api/snippet'
+  private apiUrl = 'http://localhost:8080/snippet'
 
   constructor(private http:HttpClient) { }
 
