@@ -15,7 +15,8 @@ const httpOptions = {
   providedIn: 'root'
 })
 export class TopicService {
-  private apiUrl = 'http://localhost:8080/api/topic'
+  //private apiUrl = 'http://localhost:8080/api/topic'
+  private apiUrl = 'http://localhost:8080/topic'
 
   constructor(private http:HttpClient) { }
 
